+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/api/session/room/send/SendService.kt
@@ -17,6 +17,7 @@
 package im.vector.matrix.android.api.session.room.send
 
 import im.vector.matrix.android.api.session.content.ContentAttachmentData
+import im.vector.matrix.android.api.session.events.model.Content
 import im.vector.matrix.android.api.session.events.model.Event
 import im.vector.matrix.android.api.session.room.model.message.MessageType
 import im.vector.matrix.android.api.session.room.model.message.OptionItem
@@ -27,6 +28,14 @@ import im.vector.matrix.android.api.util.Cancelable
  * This interface defines methods to send events in a room. It's implemented at the room level.
  */
 interface SendService {
+
+    /**
+     * Method to send a generic event asynchronously. If you want to send a state event, please use [StateService] instead.
+     * @param eventType the type of the event
+     * @param content the optional body as a json dict.
+     * @return a [Cancelable]
+     */
+    fun sendEvent(eventType: String, content: Content?): Cancelable
 
     /**
      * Method to send a text message asynchronously.