+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/database/mapper/AccountDataMapper.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package im.vector.matrix.android.internal.session.identity.todelete
+package im.vector.matrix.android.internal.database.mapper
 
 import com.squareup.moshi.Moshi
 import im.vector.matrix.android.api.util.JSON_DICT_PARAMETERIZED_TYPE
@@ -22,7 +22,6 @@ import im.vector.matrix.android.internal.database.model.UserAccountDataEntity
 import im.vector.matrix.android.internal.session.sync.model.accountdata.UserAccountDataEvent
 import javax.inject.Inject
 
-// There will be a duplicated class when Integration manager will be merged, so delete this one
 internal class AccountDataMapper @Inject constructor(moshi: Moshi) {
 
     private val adapter = moshi.adapter<Map<String, Any>>(JSON_DICT_PARAMETERIZED_TYPE)