+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/database/model/SessionRealmModule.kt
@@ -55,6 +55,8 @@ import io.realm.annotations.RealmModule
             HomeServerCapabilitiesEntity::class,
             RoomMemberSummaryEntity::class,
             CurrentStateEventEntity::class,
-            UserAccountDataEntity::class
+            UserAccountDataEntity::class,
+            ScalarTokenEntity::class,
+            WellknownIntegrationManagerConfigEntity::class
         ])
 internal class SessionRealmModule