+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/session/DefaultSession.kt
@@ -34,6 +34,7 @@ import im.vector.matrix.android.api.session.crypto.CryptoService
 import im.vector.matrix.android.api.session.file.FileService
 import im.vector.matrix.android.api.session.group.GroupService
 import im.vector.matrix.android.api.session.homeserver.HomeServerCapabilitiesService
+import im.vector.matrix.android.api.session.integrationmanager.IntegrationManagerService
 import im.vector.matrix.android.api.session.profile.ProfileService
 import im.vector.matrix.android.api.session.pushers.PushersService
 import im.vector.matrix.android.api.session.room.RoomDirectoryService
@@ -45,6 +46,7 @@ import im.vector.matrix.android.api.session.sync.FilterService
 import im.vector.matrix.android.api.session.sync.SyncState
 import im.vector.matrix.android.api.session.terms.TermsService
 import im.vector.matrix.android.api.session.user.UserService
+import im.vector.matrix.android.api.session.widgets.WidgetService
 import im.vector.matrix.android.internal.auth.SessionParamsStore
 import im.vector.matrix.android.internal.crypto.DefaultCryptoService
 import im.vector.matrix.android.internal.crypto.crosssigning.ShieldTrustUpdater
@@ -56,6 +58,7 @@ import im.vector.matrix.android.internal.session.room.timeline.TimelineEventDecr
 import im.vector.matrix.android.internal.session.sync.SyncTokenStore
 import im.vector.matrix.android.internal.session.sync.job.SyncThread
 import im.vector.matrix.android.internal.session.sync.job.SyncWorker
+import im.vector.matrix.android.internal.session.widgets.WidgetDependenciesHolder
 import im.vector.matrix.android.internal.task.TaskExecutor
 import im.vector.matrix.android.internal.util.MatrixCoroutineDispatchers
 import kotlinx.coroutines.Dispatchers
@@ -90,6 +93,7 @@ internal class DefaultSession @Inject constructor(
         private val fileService: Lazy<FileService>,
         private val secureStorageService: Lazy<SecureStorageService>,
         private val profileService: Lazy<ProfileService>,
+        private val widgetService: Lazy<WidgetService>,
         private val syncThreadProvider: Provider<SyncThread>,
         private val contentUrlResolver: ContentUrlResolver,
         private val syncTokenStore: SyncTokenStore,
@@ -101,11 +105,13 @@ internal class DefaultSession @Inject constructor(
         private val _sharedSecretStorageService: Lazy<SharedSecretStorageService>,
         private val accountService: Lazy<AccountService>,
         private val timelineEventDecryptor: TimelineEventDecryptor,
-        private val shieldTrustUpdater: ShieldTrustUpdater,
         private val coroutineDispatchers: MatrixCoroutineDispatchers,
         private val defaultIdentityService: DefaultIdentityService,
-        private val taskExecutor: TaskExecutor
-) : Session,
+        private val integrationManagerService: IntegrationManagerService,
+        private val taskExecutor: TaskExecutor,
+        private val widgetDependenciesHolder: WidgetDependenciesHolder,
+        private val shieldTrustUpdater: ShieldTrustUpdater)
+    : Session,
         RoomService by roomService.get(),
         RoomDirectoryService by roomDirectoryService.get(),
         GroupService by groupService.get(),
@@ -142,6 +148,7 @@ internal class DefaultSession @Inject constructor(
         timelineEventDecryptor.start()
         shieldTrustUpdater.start()
         defaultIdentityService.start()
+        widgetDependenciesHolder.start()
     }
 
     override fun requireBackgroundSync() {
@@ -187,6 +194,7 @@ internal class DefaultSession @Inject constructor(
         taskExecutor.executorScope.launch(coroutineDispatchers.main) {
             // This has to be done on main thread
             defaultIdentityService.stop()
+            widgetDependenciesHolder.stop()
         }
     }
 
@@ -232,6 +240,10 @@ internal class DefaultSession @Inject constructor(
     override fun cryptoService(): CryptoService = cryptoService.get()
 
     override fun identityService() = defaultIdentityService
+
+    override fun widgetService(): WidgetService = widgetService.get()
+
+    override fun integrationManagerService() = integrationManagerService
 
     override fun addListener(listener: Session.Listener) {
         sessionListeners.addListener(listener)