+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/RoomDetailViewState.kt
@@ -26,6 +26,7 @@ import im.vector.matrix.android.api.session.room.timeline.TimelineEvent
 import im.vector.matrix.android.api.session.sync.SyncState
 import im.vector.matrix.android.api.session.user.model.User
 import im.vector.matrix.android.api.util.MatrixItem
+import im.vector.matrix.android.api.session.widgets.model.Widget
 
 /**
  * Describes the current send mode:
@@ -55,6 +56,7 @@ data class RoomDetailViewState(
         val myRoomMember: Async<RoomMemberSummary> = Uninitialized,
         val asyncInviter: Async<User> = Uninitialized,
         val asyncRoomSummary: Async<RoomSummary> = Uninitialized,
+        val activeRoomWidgets: Async<List<Widget>> = Uninitialized,
         val typingRoomMembers: List<MatrixItem.UserItem>? = null,
         val typingMessage: String? = null,
         val sendMode: SendMode = SendMode.REGULAR(""),